refactoring the configuration environment && remove client
